Implement AdiabaticHamiltonian constructor
